fix: move CLI transcript writing to correct location
The original fix was in agent-runner-execution.ts, but CLI subagents
spawned via sessions_spawn go through commands/agent.ts instead.

Moved the transcript writing logic to run after runWithModelFallback
completes in agentCommand, which is the actual code path for CLI
subagent runs.

Tested and verified working on live server.
